LPS-93516 Move integration test UserFinderTest of com.liferay.portal.service.persistence package to module
